requirements.txt
SQLAlchemy~=2.0.40
fastapi~=0.115.12
python-dotenv~=1.1.0
structlog~=25.2.0
starlette~=0.46.2
pydantic~=2.11.3
cloudscraper~=1.2.71
requests~=2.32.3
tenacity~=9.1.2
pymemcache~=4.0.0
alembic~=1.15.2
httpx~=0.28.1
pytest==8.3.5
uvicorn==0.34.1
types-requests==2.32.0
